server/requirements.txt
openenv-core[core]>=0.2.1
fastapi>=0.115.0
uvicorn>=0.24.0
gradio>=4.0.0
plotly>=5.18.0
stable-baselines3>=2.3.0
gymnasium>=0.29.0
torch>=2.1.0
numpy>=1.24.0
openai>=1.40.0
